feat: First pass incorporating kaggle notebook.
script.py contains code that was originally written on a Kaggle
notebook. Working with this notebook became cumbersome after awhile
though, especially as I started iterating on the modeling portion.
Hence, I chunked it into functions and threw it in such a way that I
could execute it locally.

Full disclosure, this is by no means pretty, but the intent was for this
not to be pretty but to get it into a non-notebook form for iteration.
Further cleaning up is also necessary.