feat: implement models using sklearn Pipeline.
Using (https://scikit-learn.org/stable/auto_examples/ensemble/plot_stack_predictors.html#sphx-glr-auto-examples-ensemble-plot-stack-predictors-py)
as a guide, implement pipelines to do GBM, Random Forest,
Lasso, and Ridge Regression Models. Here, feature selection is done within
the bounds of cross_validation unlike the example which chooses features
ahead of time.

I tackled this as a next step because as I was thinking about fitting
more models for comparisons, I was not doing it in a good way
programmatically. I stumbled onto this example which happened to be
using the same dataset which gave a good example of how to use the
Pipeline() API. I find the ability to stamp out pipelines with
preprocessing steps in a simple manner attractive and a definite plus as I
work towards doing model validation and comparisons.
and call a single fit() method attra